Add mock-up of in page navigation bar
The intention of this mockup is to improve navigation within a page. This commit includes mock up of simple in page hyperlinks to content sections on the membership show page and the corresponding CSS. The nav.js file add sticky nav and smooth scroll features to the mockup. A similar format could be added to other show pages.